requirements.txt
playwright
streamlit
google-genai
python-dotenv
pandas
tabulate